Can switch to C interpreter and eval C code inside autology program
Really happy to prove this works.

We can now use the `with-*i*` special form to lexically bind a
predefined interpreter to our *i* interpreter environment symbol.

This involves using a custom string parser before we `read-string` the
program so that we can wrap the bodies of any `with-*i*` expressions
in strings. This allows the body to be any code we like even if it's
invalid clojure and therefore unparseable by `read-string`.