Allow interpreter modification from Python
By in-lining all the helper functions in the python interpreter we can
define it as a lambda expression data structure just like the core
Autology one.

This means we can use the core autology evaluate wrapper which allows
us to modify the *i* variable from inside python.

Needed some changes to allow the mutating `insert` list function,
ugly, but it works.